Move sum function to js directory
